add align support, split into many files
