+++ docs/pages/test-design/index.md
@@ -5,7 +5,7 @@ description: Get started with Expo in less than five minutes.
 
 import { Code, Terminal } from '~/ui/components/snippet';
 import { Select } from '~/ui/components/form';
-import { Table, Row, Cell } from '~/ui/components/table';
+import { Table, Row, Cell, TextAlign } from '~/ui/components/table';
 import { InfoIcon } from '~/ui/foundations/icons';
 
 Expo is a framework and a platform for universal React applications. It is a set of tools and services built around React Native and native platforms that help you develop, build, deploy, and quickly iterate on iOS, Android, and web apps from the same JavaScript/TypeScript codebase.
@@ -84,7 +84,10 @@ Verify that the installation was successful by running expo whoami. You're not l
 
 ### Example tables
 
-<Table headers={['Feature', 'Managed workflow', 'Bare workflow']}>
+<Table 
+  headers={['Feature', 'Managed workflow', 'Bare workflow']}
+  headersAlign={[null, undefined, TextAlign.Center]}
+>
   <Row>
     <Cell>Develop apps with only JavaScript/TypeScript</Cell>
     <Cell><InfoIcon color="green" /></Cell>
@@ -93,11 +96,11 @@ Verify that the installation was successful by running expo whoami. You're not l
   <Row>
     <Cell>Use Expo build service to create your iOS and Android builds</Cell>
     <Cell><InfoIcon color="green" /></Cell>
-    <Cell><InfoIcon color="green" /></Cell>
+    <Cell textAlign={TextAlign.Center}><InfoIcon color="green" /></Cell>
   </Row>
   <Row>
     <Cell>Develop in Xcode and Android Studio</Cell>
     <Cell></Cell>
-    <Cell><InfoIcon color="green" /></Cell>
+    <Cell textAlign={TextAlign.Center}><InfoIcon color="green" /></Cell>
   </Row>
 </Table>
