+++ docs/ui/components/table/Row.tsx
@@ -1,0 +1,21 @@
+import { css } from '@emotion/react';
+import { theme } from '@expo/styleguide';
+import React from 'react';
+
+type RowProps = {
+  children: React.ReactNode;
+};
+
+export const Row = ({ children }: RowProps) => <tr css={tableRowStyle}>{children}</tr>;
+
+const tableRowStyle = css({
+  borderBottomWidth: 1,
+  borderBottomStyle: 'solid',
+  borderBottomColor: theme.border.default,
+  '&:last-child': {
+    borderWidth: 0,
+  },
+  '&:nth-child(2n)': {
+    backgroundColor: theme.background.secondary,
+  },
+});
