+++ docs/ui/components/table/Table.shared.ts
@@ -1,0 +1,6 @@
+export enum TextAlign {
+  Left = 'left',
+  Center = 'center',
+  Right = 'right',
+  Justify = 'justify',
+}
