+++ docs/ui/components/table/index.ts
@@ -1,0 +1,4 @@
+export { Cell } from './Cell';
+export { Row } from './Row';
+export { Table } from './Table';
+export { TextAlign } from './Table.shared';
